requirements.txt
# Core dependencies
python-dotenv==1.0.1
requests==2.31.0
requests-ratelimiter==0.7.0
pyrate-limiter==3.5.0
urllib3==2.2.1
